CHALLENGE #4
Let's improve Steven's tip calculator even more, this time using loops!

Your tasks:

    Create an array called bills containing all 10 test bill values.

    Create empty arrays for the tips and the totals (tips and totals)

    Use the calcTip function we wrote before (included in the starter code) to calculate tips and total values (bill + tip) for every bill value in the bills array. Use a for loop to perform the 10 calculations!

TEST DATA: 22, 295, 176, 440, 37, 105, 10, 1100, 86, and 52.

BONUS:

Write a function calcAverage which takes an array called arr as an argument. This function calculates the average of all numbers in the given array. This is a DIFFICULT challenge (we haven't done this before)! Here is how to solve it if you feel like it: